refactor: renaming of pagination util
