+++ src/components/ui/pagination/PaginationControl.tsx
@@ -1,6 +1,6 @@
 import { Fragment } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { createPaginationArray } from './utils';
+import { createPaginationArray } from './createPaginationArray';
 
 interface PaginationControlProps {
     currentPageIndex: number;
